# docs/swagger/swagger.yaml
definitions:
...
  http.standardJsonResponse:
    properties:
      data:
        type: object
      error:
        type: string
      meta:
        type: object
      success:
        type: boolean
    type: object
...
  http.updatePasswordRequestBody:
    properties:
      password1:
        maxLength: 72
        minLength: 6
        type: string
      password2:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - password1
    - password2
    type: object
  http.userCreateRequestBody:
    properties:
      email:
        type: string
      name:
        type: string
      password1:
        maxLength: 72
        minLength: 6
        type: string
      password2:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - email
    - name
    - password1
    - password2
    type: object
...
  thunderdome.ApplicationStats:
    properties:
      activeBattleCount:
        type: integer
      activeBattleUserCount:
        type: integer
      activeRetroCount:
        type: integer
      activeRetroUserCount:
        type: integer
      activeStoryboardCount:
        type: integer
      activeStoryboardUserCount:
        type: integer
      apikeyCount:
        type: integer
      battleCount:
        type: integer
      departmentCount:
        type: integer
      estimationScaleCount:
        type: integer
      orgSubscriptionActiveCount:
        type: integer
      organizationCount:
        type: integer
      organizationEstimationScaleCount:
        type: integer
      organizationRetroTemplateCount:
        type: integer
      planCount:
        type: integer
      publicEstimationScaleCount:
        type: integer
      publicRetroTemplateCount:
        type: integer
      registeredUserCount:
        type: integer
      retroActionCount:
        type: integer
      retroCount:
        type: integer
      retroItemCount:
        type: integer
      retroTemplateCount:
        type: integer
      storyboardColumnCount:
        type: integer
      storyboardCount:
        type: integer
      storyboardGoalCount:
        type: integer
      storyboardPersonaCount:
        type: integer
      storyboardStoryCount:
        type: integer
      teamCheckinsCount:
        type: integer
      teamCount:
        type: integer
      teamEstimationScaleCount:
        type: integer
      teamRetroTemplateCount:
        type: integer
      teamSubscriptionActiveCount:
        type: integer
      unregisteredUserCount:
        type: integer
      userSubscriptionActiveCount:
        type: integer
    type: object
...
  thunderdome.Organization:
    properties:
      createdDate:
        type: string
      id:
        type: string
      name:
        type: string
      subscribed:
        type: boolean
      updatedDate:
        type: string
    type: object
...
  thunderdome.Team:
    properties:
      createdDate:
        type: string
      department_id:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
      subscribed:
        type: boolean
      updatedDate:
        type: string
    type: object
...
  thunderdome.User:
    properties:
      avatar:
        type: string
      company:
        type: string
      country:
        type: string
      createdDate:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      gravatarHash:
        type: string
      id:
        type: string
      jobTitle:
        type: string
      lastActive:
        type: string
      locale:
        type: string
      name:
        type: string
      notificationsEnabled:
        type: boolean
      picture:
        type: string
      rank:
        type: string
      theme:
        type: string
      updatedDate:
        type: string
      verified:
        type: boolean
    type: object
  thunderdome.UserAPIKey:
    properties:
      active:
        type: boolean
      apiKey:
        type: string
      createdDate:
        type: string
      id:
        type: string
      name:
        type: string
      prefix:
        type: string
      updatedDate:
        type: string
      userEmail:
        type: string
      userId:
        type: string
      userName:
        type: string
    type: object
...
info:
  contact:
    name: Steven Weathers
    url: https://github.com/StevenWeathers/thunderdome-planning-poker
  description: |-
    Thunderdome Planning Poker API for both Internal and External use.
    WARNING: Currently not considered stable and is subject to change until 1.0 is released.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thunderdome API
  version: BETA
paths:
...
  /admin/apikeys:
    get:
      description: Get a list of users API Keys
      parameters:
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.UserAPIKey'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get API Keys
      tags:
      - admin
...
  /admin/organizations:
    get:
      description: Get a list of organizations
      parameters:
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.Organization'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Organizations
      tags:
      - admin
...
  /admin/search/users/email:
    get:
      description: Get list of registered users filtered by Email likeness
      parameters:
      - description: The user Email to search for
        in: query
        name: search
        required: true
        type: string
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.User'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Search Registered Users by Email
      tags:
      - admin
  /admin/stats:
    get:
      description: Get application stats such as count of registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.ApplicationStats'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Application Stats
      tags:
      - admin
  /admin/teams:
    get:
      description: Get a list of teams
      parameters:
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.Team'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Teams
      tags:
      - admin
...
  /admin/users:
    get:
      description: Get list of registered users
      parameters:
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.User'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Registered Users
      tags:
      - admin
    post:
      description: Create a registered user
      parameters:
      - description: new user object
        in: body
        name: newUser
        required: true
        schema:
          $ref: '#/definitions/http.userCreateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Registered User
      tags:
      - admin
  /admin/users/{userId}/demote:
    patch:
      description: Demotes a user from admin to registered
      parameters:
      - description: the user ID to demote
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Demote User
      tags:
      - admin
  /admin/users/{userId}/disable:
    patch:
      description: Disable a user from logging in
      parameters:
      - description: the user ID to disable
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Disable User
      tags:
      - admin
  /admin/users/{userId}/enable:
    patch:
      description: Enable a user to allow login
      parameters:
      - description: the user ID to enable
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Enable User
      tags:
      - admin
  /admin/users/{userId}/password:
    patch:
      description: Updates the user's password
      parameters:
      - description: the user ID to update password for
        in: path
        name: userId
        required: true
        type: string
      - description: update password object
        in: body
        name: passwords
        schema:
          $ref: '#/definitions/http.updatePasswordRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Password
      tags:
      - admin
  /admin/users/{userId}/promote/:
    patch:
      description: |-
        Promotes a user to admin
        Grants read and write access to administrative information
      parameters:
      - description: the user ID to promote
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Promotes User
      tags:
      - admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"